refactoring :  프로젝트 좋아요 취소 시 접근 제어 취약점 수정

+++ src/main/java/team/themoment/everygsm/server/v2/domain/project/service/DeleteProjectLikeService.java
@@ -9,6 +9,7 @@ import org.springframework.transaction.annotation.Transactional;
 import lombok.RequiredArgsConstructor;
 import team.themoment.everygsm.server.v2.domain.project.dto.response.ProjectResDto;
 import team.themoment.everygsm.server.v2.domain.project.entity.ProjectJpaEntity;
+import team.themoment.everygsm.server.v2.domain.project.entity.constant.Status;
 import team.themoment.everygsm.server.v2.domain.project.mapper.ProjectMapper;
 import team.themoment.everygsm.server.v2.domain.project.repository.ProjectLikeRepository;
 import team.themoment.everygsm.server.v2.global.exception.error.ExpectedException;
@@ -25,6 +26,10 @@ public class DeleteProjectLikeService {
         ProjectJpaEntity project = projectLikeRepository
                 .findProjectWithCollectionsByUserIdAndProjectId(userId, projectId)
                 .orElseThrow(() -> new ExpectedException("좋아요한 프로젝트가 아닙니다.", HttpStatus.NOT_FOUND));
+
+        if (project.getStatus() != Status.APPROVED) {
+            throw new ExpectedException("승인된 프로젝트가 아닙니다.", HttpStatus.FORBIDDEN);
+        }
 
         projectLikeRepository.deleteByUserIdAndProjectId(userId, projectId);
